refactor(settings): split LiveActivitySettingsScreen into screen + widgets
846 lines holding a Hilt view-model, a screen, two dialogs and four row
widgets. Now 360, with the pieces beside it:

  LiveActivitySettingsViewModel.kt   118  state holder + State
  LiveActivityCustomTimeDialogs.kt   245  the two 自訂 entry sheets
  LiveActivityPermissionRow.kt       100  status dots + tap routing
  LiveActivityLeadTimeRow.kt          91  slider row + duration label

Pure move — every relocated line is byte-identical to its original,
verified by diffing the reassembled members against `git show HEAD:`.
Kotlin `private` is file-scoped, so the extracted members widen to
`internal`; nothing outside this package can see them either way.

Two deliberate deviations, both to avoid leaving a trap behind:

`formatDuration` became `formatLeadDuration`. It was private and could
afford a vague name; as an `internal` member of a package that already
holds fifteen settings screens, "format a duration" is a name the next
person would reasonably reuse for something else.

`SettingToggleRow` is gone; its eight call sites now use the existing
`SettingsToggleRow` from SettingsRows.kt. Extracting it as-is would have
put `SettingToggleRow` and `SettingsToggleRow` — one character apart —
in the same package as two internal composables, which is worse than
the duplication was when one of them was private. The two rendered
identically: `Color.Unspecified` resolves to LocalContentColor, which
inside ContentCard is the same `onSurface` the shared row names
explicitly, and both disable to `onSurface * ContentAlpha.DISABLED`.
The one behavioural difference is an improvement — the shared row is
`heightIn(min: 56.dp)` where the old one was a fixed `height(56.dp)`,
so a label that wraps to two lines in some locale now grows the row
instead of being clipped by it.

:app:compilePlayDebugKotlin, :app:compileFdroidDebugKotlin,
:wear:compileDebugKotlin all clean; 196 unit tests pass.
